MODIFY: SearchBar(검색창) 분리 및 스타일링 조정
