+++ src/components/Nav/Nav.js
@@ -1,10 +1,15 @@
 import React, { useState } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import Categories from './Categories';
+import SearchBar from './SearchBar';
 import './Nav.scss';
 
 const Nav = () => {
   const [showSearchBar, setShowSearchBar] = useState(false);
+  const navigate = useNavigate();
+  const goToLogin = () => {
+    navigate('/login');
+  };
   const showBar = () => {
     setShowSearchBar(prev => !prev);
   };
@@ -22,9 +27,10 @@ const Nav = () => {
         <Categories />
       </ul>
       <div className="navIcons">
-        <Link to="#">
+        <span className="loginButton" onClick={goToLogin}>
+          {/* TODO: 추후 Login페이지 이동에 관련 로그인 되어있을 때와 아닐때 로직 작성 요망 */}
           <div className="iconProfile" />
-        </Link>
+        </span>
         <Link to="#">
           <div className="iconLocation" />
         </Link>
@@ -38,41 +44,5 @@ const Nav = () => {
 };
 
 // 작업 완료되는 대로 분리 필요
-const SearchBar = () => {
-  return (
-    <div className="searchPad">
-      <div className="search">
-        <div className="searchWrap">
-          <form className="searchForm">
-            <div className="searchBox">
-              <span className="inputSearchIcon" />
-              <input
-                type="search"
-                placeholder="검색어를 입력해 주세요"
-                className="searchInput"
-              />
-            </div>
-            <button className="searchButton">검색</button>
-          </form>
-          <div className="searchRecommends">
-            <span>추천 검색어</span>
-            <Link to="#" className="tags">
-              볼륨 마스카라
-            </Link>
-            <Link to="#" className="tags">
-              소프트 네츄럴 립스틱
-            </Link>
-            <Link to="#" className="tags">
-              하이드레이팅 프라이머
-            </Link>
-            <Link to="#" className="tags">
-              시그니처 브로우 펜슬
-            </Link>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 export default Nav;
