+++ src/pages/Main/Main.js
@@ -1,8 +1,14 @@
 import React from 'react';
+import Nav from '../../components/Nav/Nav';
 import './Main.scss';
 
 const Main = () => {
-  return <div>메인</div>;
+  return (
+    <>
+      <Nav />
+      <div>메인</div>
+    </>
+  );
 };
 
 export default Main;
